jobs/requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
cachetools==5.3.0
celery==5.2.7
celery-pubsub==1.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
docopt==0.6.2
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
google-api-core==2.11.0
google-auth==2.18.1
google-cloud-core==2.3.2
google-cloud-pubsub==2.17.0
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
pipreqs==0.4.13
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytz==2023.3
redis==4.5.4
requests==2.28.2
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
yarg==0.1.9
